Revert "update prefixed schema as well"
This reverts commit b0ec92dad6184e9e1d00147943bd5a8a5beef6ee.

+++ SQL/paradise_schema_prefixed.sql
@@ -253,7 +253,7 @@ CREATE TABLE `SS13_player` (
   `UI_style_alpha` smallint(4) DEFAULT '255',
   `be_role` mediumtext,
   `default_slot` smallint(4) DEFAULT '1',
-  `toggles` int(8) DEFAULT '383',
+  `toggles` mediumint(8) DEFAULT '383',
   `sound` mediumint(8) DEFAULT '31',
   `randomslot` tinyint(1) DEFAULT '0',
   `volume` smallint(4) DEFAULT '100',
